Token based authentication and authorization

+++ shopping-list/src/main/resources/db/migration/V0.6__Add_Verification_Request_Table.sql
@@ -3,6 +3,6 @@ create table VERIFICATION_REQUEST
   ID                UUID PRIMARY KEY,
   USER_ID           UUID,
   VERIFICATION_CODE TEXT,
-  VALIDITY          INT,
+  VALIDITY          BIGINT,
   CREATED_DATE      TIMESTAMP
 )
